requirements.txt
PyYAML>=6.0
httpx>=0.27.0
qdrant-client>=1.9.0
watchdog>=4.0.0
